+++ crate-plugin/src/main/java/me/oriley/crate/CrateGenerator.java
@@ -26,8 +26,10 @@ import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.imageio.ImageIO;
 import javax.lang.model.element.Modifier;
 import java.awt.*;
+import java.awt.image.BufferedImage;
 import java.io.*;
 import java.util.*;
 import java.util.List;
@@ -36,7 +38,9 @@ import java.util.concurrent.TimeUnit;
 import static java.util.Locale.US;
 import static javax.lang.model.element.Modifier.*;
 import static me.oriley.crate.utils.JavaPoetUtils.*;
-import static org.apache.commons.lang.StringUtils.equalsIgnoreCase;
+import static me.oriley.crate.utils.JavaPoetUtils.Nullability.NONE;
+import static me.oriley.crate.utils.JavaPoetUtils.Nullability.NONNULL;
+import static me.oriley.crate.utils.JavaPoetUtils.Nullability.NULLABLE;
 
 public final class CrateGenerator {
 
@@ -44,21 +48,30 @@ public final class CrateGenerator {
     private static final String CRATE_HASH = CrateHasher.getActualHash();
     private static final String ASSETS = "assets";
     private static final String CONTEXT = "context";
+    private static final String BITMAP = "bitmap";
     private static final String TYPEFACE = "typeface";
     private static final String OPEN = "open";
+    private static final String GET_BITMAP = "getBitmap";
     private static final String GET_TYPEFACE = "getTypeface";
     private static final String THIS = "this";
+    private static final String CLEAR = "clear";
+    private static final String STREAM = "stream";
+    private static final String KEY = "key";
+    private static final String DEBUG = "DEBUG";
 
+    private static final String BITMAP_CACHE = "bitmapCache";
     private static final String TYPEFACE_CACHE = "typefaceCache";
     private static final String ASSET_MANAGER = "AssetManager";
 
+    private static final ClassName BITMAP_FACTORY_CLASS = ClassName.get("android.graphics", "BitmapFactory");
     private static final ClassName LOG_CLASS = ClassName.get("android.util", "Log");
     private static final ClassName ASSETMANAGER_CLASS = ClassName.get("android.content.res", ASSET_MANAGER);
+    private static final ClassName BITMAP_CLASS = ClassName.get("android.graphics", capitalise(BITMAP));
     private static final ClassName CONTEXT_CLASS = ClassName.get("android.content", capitalise(CONTEXT));
     private static final ClassName TYPEFACE_CLASS = ClassName.get("android.graphics", capitalise(TYPEFACE));
 
-    private static final String OTF_EXTENSION = "otf";
-    private static final String TTF_EXTENSION = "ttf";
+    private static final List<String> FONT_EXTENSIONS = Arrays.asList("otf", "ttf");
+    private static final List<String> IMAGE_EXTENSIONS = Arrays.asList("jpg", "jpeg", "gif", "png");
 
     private static final Logger log = LoggerFactory.getLogger(CrateGenerator.class.getSimpleName());
 
@@ -197,10 +210,13 @@ public final class CrateGenerator {
                 .addModifiers(PUBLIC, FINAL)
                 .addType(createAssetClass())
                 .addType(createFontAssetClass())
+                .addType(createImageAssetClass())
                 .addAnnotation(createSuppressWarningAnnotation("unused"));
 
-        builder.addField(createField(ASSETMANAGER_CLASS, false, PRIVATE, FINAL))
-                .addField(createTypefaceCacheField());
+        builder.addField(createBooleanField(DEBUG, mDebugLogging, PRIVATE, STATIC, FINAL))
+                .addField(createField(ASSETMANAGER_CLASS, NONNULL, PRIVATE, FINAL))
+                .addField(createCacheMapField(TYPEFACE_CACHE, TYPEFACE_CLASS))
+                .addField(createCacheMapField(BITMAP_CACHE, BITMAP_CLASS));
 
         TreeMap<String, Asset> allAssets = new TreeMap<>();
         listFiles(allAssets, builder, ASSETS, variantDir, variantAssetDir, true);
@@ -214,7 +230,9 @@ public final class CrateGenerator {
         builder.addMethod(createCrateConstructor())
                 .addMethod(createInputStreamMethod(false))
                 .addMethod(createInputStreamMethod(true))
-                .addMethod(createTypefaceMethod());
+                .addMethod(createBitmapMethod())
+                .addMethod(createTypefaceMethod())
+                .addMethod(createBitmapClearMethod())
 
         JavaFile.Builder javaBuilder = JavaFile.builder(packageName, builder.build())
                 .indent("    ");
@@ -249,6 +267,7 @@ public final class CrateGenerator {
         List<File> files = getFileList(directory);
         TreeMap<String, Asset> assetMap = new TreeMap<>();
         boolean isFontFolder = true;
+        boolean isImageFolder = true;
 
         for (File file : files) {
             if (file.isDirectory()) {
@@ -267,14 +286,33 @@ public final class CrateGenerator {
 
                 String filePath = file.getPath().replace(variantAssetDir + "/", "");
 
-                String fileExtension = getFileExtension(fileName);
+                String fileExtension = getFileExtension(fileName).toLowerCase(US);
                 Asset asset;
-                if (equalsIgnoreCase(fileExtension, TTF_EXTENSION) || equalsIgnoreCase(fileExtension, OTF_EXTENSION)) {
+                if (FONT_EXTENSIONS.contains(fileExtension)) {
+                    isImageFolder = false;
                     String fontName = getFontName(file.getPath());
                     asset = new FontAsset(fieldName, filePath, fileName, fontName != null ? fontName : fileName);
                     builder.addField(createFontAssetField((FontAsset) asset));
+                } else if (IMAGE_EXTENSIONS.contains(fileExtension)) {
+                    isFontFolder = false;
+
+                    int width = 0;
+                    int height = 0;
+                    try {
+                        BufferedImage image = ImageIO.read(file);
+                        if (image != null) {
+                            width = image.getWidth();
+                            height = image.getHeight();
+                        }
+                    } catch (IOException e) {
+                        logError("Error parsing image: " + file.getPath(), e, false);
+                    }
+
+                    asset = new ImageAsset(fieldName, filePath, fileName, width, height);
+                    builder.addField(createImageAssetField((ImageAsset) asset));
                 } else {
                     isFontFolder = false;
+                    isImageFolder = false;
                     asset = new Asset(fieldName, filePath, fileName);
                     builder.addField(createAssetField(asset));
                 }
@@ -284,7 +322,8 @@ public final class CrateGenerator {
         }
 
         if (!assetMap.isEmpty()) {
-            TypeName elementType = TypeVariableName.get(isFontFolder ? FontAsset.getTypeName() : Asset.getTypeName());
+            TypeName elementType = TypeVariableName.get(isFontFolder ? FontAsset.getTypeName() :
+                    isImageFolder ? ImageAsset.getTypeName() : Asset.getTypeName());
             TypeName listType = ParameterizedTypeName.get(ClassName.get(List.class), elementType);
             builder.addField(createListField(listType, "LIST", assetMap));
         }
@@ -320,11 +359,19 @@ public final class CrateGenerator {
         }
 
         for (String field : fields) {
-            builder.addField(createStringField(toInstance(field), false, FINAL))
-                    .addMethod(createGetter(field, String.class, false, PUBLIC));
+            builder.addField(createStringField(field, NONNULL, FINAL))
+                    .addMethod(createGetter(field, String.class, NONNULL, PUBLIC));
         }
 
         return builder.build();
+    }
+
+    private void addAssetSuperConstructor(@NonNull MethodSpec.Builder builder) {
+        String[] fields = Asset.getFields();
+        builder.addStatement("super(" + Joiner.on(", ").join(fields) + ")");
+        for (String field : fields) {
+            builder.addParameter(createParameter(field, String.class, NONNULL));
+        }
     }
 
     @NonNull
@@ -332,10 +379,10 @@ public final class CrateGenerator {
         validateNonStatic();
 
         MethodSpec.Builder builder = MethodSpec.methodBuilder(OPEN)
-                .addAnnotation(createNullabilityAnnotation(false))
                 .addModifiers(PUBLIC)
                 .addException(IOException.class)
                 .returns(InputStream.class);
+        addNullability(builder, NONNULL);
 
         String modeName = "mode";
         if (mode) {
@@ -362,32 +409,92 @@ public final class CrateGenerator {
 
         String[] fields = FontAsset.getFields();
         for (String field : fields) {
-            builder.addField(createStringField(toInstance(field), false, FINAL))
-                    .addMethod(createGetter(field, String.class, false, PUBLIC));
+            builder.addField(createStringField(field, NONNULL, FINAL))
+                    .addMethod(createGetter(field, String.class, NONNULL, PUBLIC));
         }
 
-        builder.addMethod(createSubConstructor(Asset.getFields(), fields));
+        builder.addMethod(createFontAssetConstructor());
+        return builder.build();
+    }
+
+    @NonNull
+    private MethodSpec createFontAssetConstructor() {
+        MethodSpec.Builder builder = MethodSpec.constructorBuilder()
+                .addModifiers(Modifier.PRIVATE);
+        addAssetSuperConstructor(builder);
+
+        for (String field : FontAsset.getFields()) {
+            builder.addParameter(createParameter(field, String.class, NONNULL))
+                    .addStatement("$N = $N", asFieldName(field), field);
+        }
+
         return builder.build();
     }
 
     @NonNull
     private MethodSpec createNonStaticTypefaceMethod() {
         validateNonStatic();
-
-        return MethodSpec.methodBuilder(GET_TYPEFACE)
-                .addAnnotation(createNullabilityAnnotation(true))
+        MethodSpec.Builder builder = MethodSpec.methodBuilder(GET_TYPEFACE)
                 .addModifiers(PUBLIC)
                 .addStatement("return $N.$N.$N($N)", mClassName, THIS, GET_TYPEFACE, THIS)
-                .returns(TYPEFACE_CLASS)
-                .build();
+                .returns(TYPEFACE_CLASS);
+        addNullability(builder, NULLABLE);
+        return builder.build();
+    }
+
+    @NonNull
+    private TypeSpec createImageAssetClass() {
+        TypeSpec.Builder builder = TypeSpec.classBuilder(ImageAsset.getTypeName())
+                .superclass(TypeVariableName.get(Asset.getTypeName()))
+                .addModifiers(PUBLIC);
+
+        if (mStaticMode) {
+            builder.addModifiers(STATIC);
+        } else {
+            builder.addMethod(createNonStaticBitmapMethod());
+        }
+
+        String[] fields = ImageAsset.getFields();
+        for (String field : fields) {
+            builder.addField(createIntField(field, FINAL))
+                    .addMethod(createGetter(field, int.class, NONE, PUBLIC));
+        }
+
+        builder.addMethod(createImageAssetConstructor());
+        return builder.build();
+    }
+
+    @NonNull
+    private MethodSpec createNonStaticBitmapMethod() {
+        validateNonStatic();
+        MethodSpec.Builder builder = MethodSpec.methodBuilder(GET_BITMAP)
+                .addModifiers(PUBLIC)
+                .addStatement("return $N.$N.$N($N)", mClassName, THIS, GET_BITMAP, THIS)
+                .returns(BITMAP_CLASS);
+        addNullability(builder, NULLABLE);
+        return builder.build();
+    }
+
+    @NonNull
+    private MethodSpec createImageAssetConstructor() {
+        MethodSpec.Builder builder = MethodSpec.constructorBuilder()
+                .addModifiers(Modifier.PRIVATE);
+        addAssetSuperConstructor(builder);
+
+        for (String field : ImageAsset.getFields()) {
+            builder.addParameter(createParameter(field, int.class, NONE))
+                    .addStatement("$N = $N", asFieldName(field), field);
+        }
+
+        return builder.build();
     }
 
     @NonNull
     private MethodSpec createCrateConstructor() {
         return MethodSpec.constructorBuilder()
                 .addModifiers(Modifier.PUBLIC)
-                .addParameter(createParameter(CONTEXT, CONTEXT_CLASS, false))
-                .addStatement("$N = $N.getApplicationContext().getAssets()", toInstance(ASSET_MANAGER), CONTEXT)
+                .addParameter(createParameter(CONTEXT, CONTEXT_CLASS, NONNULL))
+                .addStatement("$N = $N.getApplicationContext().getAssets()", asFieldName(ASSET_MANAGER), CONTEXT)
                 .build();
     }
 
@@ -397,53 +504,97 @@ public final class CrateGenerator {
                 .addModifiers(Modifier.PRIVATE);
 
         for (String field : fields) {
-            builder.addParameter(createParameter(field, String.class, false))
-                    .addStatement("$N = $N", toInstance(field), field);
+            builder.addParameter(createParameter(field, String.class, NONNULL))
+                    .addStatement("$N = $N", asFieldName(field), field);
         }
 
         return builder.build();
     }
 
-    @NonNull
-    private FieldSpec createTypefaceCacheField() {
-        TypeName mapType = ParameterizedTypeName.get(ClassName.get(HashMap.class), ClassName.get(String.class), TYPEFACE_CLASS);
-        return FieldSpec.builder(mapType, toInstance(TYPEFACE_CACHE))
+    private FieldSpec createCacheMapField(@NonNull String name, @NonNull ClassName entryClass) {
+        TypeName mapType = ParameterizedTypeName.get(ClassName.get(HashMap.class), ClassName.get(String.class), entryClass);
+        FieldSpec.Builder builder = FieldSpec.builder(mapType, asFieldName(name))
                 .addModifiers(PRIVATE, FINAL)
-                .addAnnotation(createNullabilityAnnotation(false))
-                .initializer(CodeBlock.builder().add("new $T()", mapType).build())
-                .build();
+                .initializer(CodeBlock.builder().add("new $T()", mapType).build());
+        addNullability(builder, NONNULL);
+        return builder.build();
+    }
+
+    @NonNull
+    private MethodSpec createBitmapMethod() {
+        String imageAsset = "imageAsset";
+        String cacheField = asFieldName(BITMAP_CACHE);
+
+        TypeName assetType = TypeVariableName.get(ImageAsset.getTypeName());
+        MethodSpec.Builder builder = MethodSpec.methodBuilder(GET_BITMAP)
+                .addModifiers(mStaticMode ? PUBLIC : PRIVATE)
+                .addParameter(createParameter(imageAsset, assetType, NONNULL))
+                .addCode(CodeBlock.builder()
+                        .addStatement("$T $N = $N.mPath", String.class, KEY, imageAsset)
+                        .beginControlFlow("if ($N.containsKey($N))", cacheField, KEY)
+                        .addStatement("$T $N = $N.get($N)", BITMAP_CLASS, BITMAP, cacheField, KEY)
+                        .beginControlFlow("if ($N != null && !$N.isRecycled())", BITMAP, BITMAP)
+                        .addStatement("if ($N) $T.d($S, \"Using cached bitmap for key: \" + $N)", DEBUG, LOG_CLASS, mClassName, KEY)
+                        .addStatement("return $N", BITMAP)
+                        .nextControlFlow("else")
+                        .addStatement("if ($N) $T.d($S, \"Ejecting recycled bitmap for key: \" + $N)", DEBUG, LOG_CLASS, mClassName, KEY)
+                        .addStatement("$N.remove($N)", cacheField, KEY)
+                        .endControlFlow()
+                        .endControlFlow()
+                        .addStatement("$T $N = null", BITMAP_CLASS, BITMAP)
+                        .beginControlFlow("try")
+                        .addStatement("$T $N = open($N)", InputStream.class, STREAM, imageAsset)
+                        .beginControlFlow("try")
+                        .addStatement("$N = $T.decodeStream($N)", BITMAP, BITMAP_FACTORY_CLASS, STREAM)
+                        .nextControlFlow("finally")
+                        .addStatement("$N.close()", STREAM)
+                        .endControlFlow()
+                        .nextControlFlow("catch ($T e)", IOException.class)
+                        .addStatement("$T.e($S, \"Failed to load bitmap for key: \" + $N, e)", LOG_CLASS, mClassName, KEY)
+                        .addStatement("e.printStackTrace()")
+                        .nextControlFlow("finally")
+                        .beginControlFlow("if ($N != null)", BITMAP)
+                        .addStatement("if ($N) $T.d($S, \"Bitmap loaded for key: \" + $N)", DEBUG, LOG_CLASS, mClassName, KEY)
+                        .addStatement("$N.put($N, $N)", cacheField, KEY, BITMAP)
+                        .endControlFlow()
+                        .addStatement("return $N", BITMAP)
+                        .endControlFlow()
+                        .build())
+                .returns(BITMAP_CLASS);
+        addNullability(builder, NULLABLE);
+        return builder.build();
     }
 
     @NonNull
     private MethodSpec createTypefaceMethod() {
         String fontAsset = "fontAsset";
-        String localVarKey = "key";
+        String cacheField = asFieldName(TYPEFACE_CACHE);
 
         TypeName assetType = TypeVariableName.get(FontAsset.getTypeName());
-        return MethodSpec.methodBuilder(GET_TYPEFACE)
-                .addAnnotation(createNullabilityAnnotation(true))
+        MethodSpec.Builder builder = MethodSpec.methodBuilder(GET_TYPEFACE)
                 .addModifiers(mStaticMode ? PUBLIC : PRIVATE)
-                .addParameter(createParameter(fontAsset, assetType, false))
+                .addParameter(createParameter(fontAsset, assetType, NONNULL))
                 .addCode(CodeBlock.builder()
-                        .addStatement("$T $N = $N.mPath", String.class, localVarKey, fontAsset)
-                        .beginControlFlow("if ($N.containsKey($N))", toInstance(TYPEFACE_CACHE), localVarKey)
-                        .addStatement("return $N.get($N)", toInstance(TYPEFACE_CACHE), localVarKey)
+                        .addStatement("$T $N = $N.mPath", String.class, KEY, fontAsset)
+                        .beginControlFlow("if ($N.containsKey($N))", cacheField, KEY)
+                        .addStatement("return $N.get($N)", cacheField, KEY)
                         .endControlFlow()
                         .addStatement("$T $N = null", TYPEFACE_CLASS, TYPEFACE)
                         .beginControlFlow("try")
-                        .addStatement("$N = $T.createFromAsset($N, $N.mPath)", TYPEFACE, TYPEFACE_CLASS, toInstance(ASSET_MANAGER), fontAsset)
+                        .addStatement("$N = $T.createFromAsset($N, $N)", TYPEFACE, TYPEFACE_CLASS, asFieldName(ASSET_MANAGER), KEY)
                         .nextControlFlow("catch ($T e)", RuntimeException.class)
-                        .addStatement("$T.e($S, \"Failed to load typeface for key: \" + $N, e)", LOG_CLASS, mClassName, localVarKey)
+                        .addStatement("$T.e($S, \"Failed to load typeface for key: \" + $N, e)", LOG_CLASS, mClassName, KEY)
                         .addStatement("e.printStackTrace()")
                         .nextControlFlow("finally")
                         .beginControlFlow("if ($N != null)", TYPEFACE)
-                        .addStatement("$N.put($N, $N)", toInstance(TYPEFACE_CACHE), localVarKey, TYPEFACE)
-                        .endControlFlow()
+                        .addStatement("$N.put($N, $N)", cacheField, KEY, TYPEFACE)
                         .endControlFlow()
                         .addStatement("return $N", TYPEFACE)
+                        .endControlFlow()
                         .build())
-                .returns(TYPEFACE_CLASS)
-                .build();
+                .returns(TYPEFACE_CLASS);
+        addNullability(builder, NULLABLE);
+        return builder.build();
     }
 
     @NonNull
@@ -451,40 +602,32 @@ public final class CrateGenerator {
         String asset = "asset";
         TypeName assetType = TypeVariableName.get(Asset.getTypeName());
         MethodSpec.Builder builder = MethodSpec.methodBuilder(OPEN)
-                .addAnnotation(createNullabilityAnnotation(false))
                 .addModifiers(mStaticMode ? PUBLIC : PRIVATE)
-                .addParameter(createParameter("asset", assetType, false))
+                .addParameter(createParameter("asset", assetType, NONNULL))
                 .addException(IOException.class)
                 .returns(InputStream.class);
+        addNullability(builder, NONNULL);
 
         String modeName = "mode";
         if (mode) {
             builder.addParameter(createPrimitiveParameter(modeName, int.class))
-                    .addStatement("return $N.open($N.mPath, $N)", toInstance(ASSET_MANAGER), asset, modeName);
+                    .addStatement("return $N.open($N.mPath, $N)", asFieldName(ASSET_MANAGER), asset, modeName);
         } else {
-            builder.addStatement("return $N.open($N.mPath)", toInstance(ASSET_MANAGER), asset);
+            builder.addStatement("return $N.open($N.mPath)", asFieldName(ASSET_MANAGER), asset);
         }
 
         return builder.build();
     }
 
     @NonNull
-    private MethodSpec createSubConstructor(@NonNull String[] parentFields, @NonNull String[] fields) {
-        MethodSpec.Builder builder = MethodSpec.constructorBuilder()
-                .addModifiers(Modifier.PRIVATE);
-
-        builder.addStatement("super(" + Joiner.on(", ").join(parentFields) + ")");
-
-        for (String field : parentFields) {
-            builder.addParameter(createParameter(field, String.class, false));
-        }
-
-        for (String field : fields) {
-            builder.addParameter(createParameter(field, String.class, false))
-                    .addStatement("$N = $N", toInstance(field), field);
-        }
-
-        return builder.build();
+    private MethodSpec createBitmapClearMethod() {
+        return MethodSpec.methodBuilder(CLEAR + capitalise(BITMAP_CACHE))
+                .addModifiers(PUBLIC)
+                .beginControlFlow("for ($T $N : $N.values())", BITMAP_CLASS, BITMAP, asFieldName(BITMAP_CACHE))
+                .addStatement("$N.recycle()", BITMAP)
+                .endControlFlow()
+                .addStatement("$N.clear()", asFieldName(BITMAP_CACHE))
+                .build();
     }
 
     @NonNull
@@ -516,11 +659,11 @@ public final class CrateGenerator {
         validateNonStatic();
 
         TypeName typeName = TypeVariableName.get(makeClassName(rootName));
-        return FieldSpec.builder(typeName, rootName)
+        FieldSpec.Builder builder = FieldSpec.builder(typeName, rootName)
                 .addModifiers(PUBLIC, FINAL)
-                .addAnnotation(createNullabilityAnnotation(false))
-                .initializer("new $T()", typeName)
-                .build();
+                .initializer("new $T()", typeName);
+        addNullability(builder, NONNULL);
+        return builder.build();
     }
 
     @NonNull
@@ -539,6 +682,19 @@ public final class CrateGenerator {
     @NonNull
     private FieldSpec createFontAssetField(@NonNull FontAsset asset) {
         FieldSpec.Builder builder = FieldSpec.builder(TypeVariableName.get(FontAsset.getTypeName()), asset.getFieldName())
+                .addModifiers(PUBLIC, FINAL);
+
+        if (mStaticMode) {
+            builder.addModifiers(STATIC);
+        }
+
+        asset.addInitialiser(builder);
+        return builder.build();
+    }
+
+    @NonNull
+    private FieldSpec createImageAssetField(@NonNull ImageAsset asset) {
+        FieldSpec.Builder builder = FieldSpec.builder(TypeVariableName.get(ImageAsset.getTypeName()), asset.getFieldName())
                 .addModifiers(PUBLIC, FINAL);
 
         if (mStaticMode) {
@@ -698,6 +854,46 @@ public final class CrateGenerator {
         @NonNull
         public static String[] getFields() {
             return new String[]{"fontName"};
+        }
+    }
+
+    @SuppressWarnings("unused")
+    private static final class ImageAsset extends Asset {
+
+        final int mWidth;
+
+        final int mHeight;
+
+        private ImageAsset(@NonNull String fieldName,
+                           @NonNull String path,
+                           @NonNull String name,
+                           int width,
+                           int height) {
+            super(fieldName, path, name);
+            mWidth = width;
+            mHeight = height;
+        }
+
+        public int getWidth() {
+            return mWidth;
+        }
+
+        public int getHeight() {
+            return mHeight;
+        }
+
+        public void addInitialiser(@NonNull FieldSpec.Builder builder) {
+            builder.initializer("new $N($S, $S, $L, $L)", getTypeName(), mPath, mName, mWidth, mHeight);
+        }
+
+        @NonNull
+        public static String getTypeName() {
+            return ImageAsset.class.getSimpleName();
+        }
+
+        @NonNull
+        public static String[] getFields() {
+            return new String[]{"width", "height"};
         }
     }
 }