+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Change Log
 ==========
 
+## Version 0.3 WIP
+
+ *  Feature: Add ImageAsset class for JPG/JPEG/PNG/GIF assets, with width and height fields
+ *  Feature: Implement convenience getter for ImageAsset bitmaps, and bitmap caching
+
+
 ## Version 0.2
 
 _2016-3-14